Refactor main.py to allow importing billing data from multiple platforms
The main.py file has been refactored to allow the user to choose which platform's billing data they want to import, or import data from all platforms. The user is prompted to enter a number (0 for wechatpay, 1 for alipay, 2 for all) and the corresponding billing data is imported to Notion. This change improves the flexibility and usability of the application.

A new file main_without_mail.py has also been added, which contains the same functionality as main.py but without the email integration. This allows users who don't need the email functionality to use a lighter version of the application.